refactor(tui): organize station button hooks
